backend/requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
drf-extra-fields==3.5.0
isort==5.12.0 
filetype==1.2.0
Pillow==9.5.0
social-auth-app-django==4.0.0
webcolors==1.13